Simple shopping cart
It is a java code which displays the list of items. The user has to choose the item number and add the quantity. The code will show the total price of the items. 

Motivation
To reduce human effort and time consumption in calculating the total amount. 

Tech/Framework
Built with Netbeans IDE 8.2

Collaborators
@R-Kanguri
@gmadhur
@dorkadiriya